--- src/presentation/controllers/login/login-controller-protocols.ts
@@ -1,2 +1,0 @@
-export * from '../../protocols'
-export * from '../../../domain/usecases/authentication'
