+++ src/presentation/controllers/login/login/login-controller-protocols.ts
@@ -1,0 +1,2 @@
+export * from '../../../protocols'
+export * from '../../../../domain/usecases/authentication'
