--- src/presentation/controllers/signup/signup-controller-protocols.ts
@@ -1,4 +1,0 @@
-export * from '../../protocols'
-export * from '../../../domain/usecases/add-account'
-export * from '../../../domain/usecases/authentication'
-export * from '../../../domain/models/account'
